Reg form dir change
